Split a foundation mod from appkit.
